fix: [김혜인] 수정 (및 feedback) [DiceGame_2-5]

+++ dicegame/src/account/repository/UserRepository.java
@@ -7,5 +7,6 @@ import java.util.Optional;
 public interface UserRepository {
     int save(User user);
     Optional<User> findByUserId(String userId);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
+
 }
